rename article-promo => article-stub
