+++ server/model/article-stub.js
